Trait oak_functions_service::logger::OakLogger

pub trait OakLogger: Send + Sync {
    // Required methods
    fn log_sensitive(&self, level: Level, message: &str);
    fn log_public(&self, level: Level, message: &str);
}

Required Methods

fn log_sensitive(&self, level: Level, message: &str)

Logs the message, which might contain sensitive information, at the specified Level.

Only insecure debug-only implementations may provide a non-empty implementation. Production implementations must not do anything.

fn log_public(&self, level: Level, message: &str)

Logs a message that contains only public, non-sensitive content at the specified Level.

All code that uses this function must be inspected to ensure that the message can never contain any information that could have been derived from sensitive or non-public data.