Struct oak_functions_service::lookup::LookupDataManager

pub struct LookupDataManager<L: OakLogger + Clone> { /* private fields */ }

Utility for managing lookup data.

LookupDataManager can be used to create LookupData instances that share the underlying data. It can also update the underlying data. After updating the data, new LookupData instances will use the new data, but earlier instances will still used the earlier data.

LookupDataManager maintains the invariants consistent view on lookup data , and shared lookup data

Note that the data is never mutated in-place, but only ever replaced. So instead of the Rust idiom Arc<Spinlock<T>> we have Spinlock<Arc<T>>.

In the future we may replace both the mutex and the hash map with something like RCU.

Implementations

impl<L> LookupDataManager<L>where L: OakLogger + Clone,

pub fn new_empty(logger: L) -> Self

Creates a new instance with empty backing data.

pub fn for_test(data: Data, logger: L) -> Self

Creates an instance of LookupData populated with the given entries.

pub fn extend_next_lookup_data(&self, new_data: Data)

pub fn finish_next_lookup_data(&self)

pub fn abort_next_lookup_data(&self)

pub fn create_lookup_data(&self) -> LookupData<L>

Creates a new LookupData instance with a reference to the current backing data.