Module oak_functions_service::schema

Structs

Empty

InitializeRequest

InitializeResponse

InvokeRequest

InvokeResponse

LookupData

If the definition of LookupData changes, the estimation of the size of the LookupData when serialized in the Oak Functions Launcher needs to change, too.

LookupDataEntry

OakFunctionsAsyncClient

OakFunctionsClient

OakFunctionsServer

PublicKeyInfo

Traits

OakFunctions