Trait oak_functions_service::Handler

pub trait Handler {
    type HandlerType: Handler;
    type HandlerConfig: Default + Send + Sync + Clone;

    // Required methods
    fn new_handler<const S: usize>(
        config: Self::HandlerConfig,
        wasm_module_bytes: &[u8],
        lookup_data_manager: Arc<LookupDataManager<S>>,
        observer: Option<Arc<dyn Observer + Send + Sync>>,
    ) -> Result<Self::HandlerType>;
    fn handle_invoke(&self, invoke_request: Request) -> Result<Response, Status>;
}

Required Associated Types

type HandlerType: Handler

type HandlerConfig: Default + Send + Sync + Clone

Required Methods

fn new_handler<const S: usize>( config: Self::HandlerConfig, wasm_module_bytes: &[u8], lookup_data_manager: Arc<LookupDataManager<S>>, observer: Option<Arc<dyn Observer + Send + Sync>>, ) -> Result<Self::HandlerType>

fn handle_invoke(&self, invoke_request: Request) -> Result<Response, Status>

Handles a call to invoke by getting the raw request bytes from the body of the request to invoke and returns a reponse to invoke setting the raw bytes in the body of the response.

Object Safety

This trait is not object safe.