Module oak_functions_service::wasm

Wasm business logic provider based on Wasmi.

Modules

api

wasmtime

Wasm business logic provider based on wasmtime.

Structs

UserState

UserState holds the user request bytes and response bytes for a particular execution of an Oak Wasm module. The UserState also holds a reference to the logger and the enabled extensions.

WasmConfig

WasmHandler

Constants

ALLOC_FUNCTION_NAME

Fixed name of the function to allocate memory. Every Oak Wasm module must provide this function.

MAIN_FUNCTION_NAME

Fixed name of the function to start a Wasm. Every Oak Wasm module must provide this function.

MEMORY_NAME

The name of the memory every Oak Wasm module has.

Traits

WasmApi

A trait for Wasm APIs that can be called from Wasm modules.

WasmApiFactory

A trait for creating Wasm APIs that can be called from Wasm modules.

Type Aliases

AbiPointer

Type aliases for positions in Wasm linear memory. Any future 64-bit version of Wasm would use different types.

AbiPointerOffset

Type aliases for offsets in Wasm linear memory.