Struct oak_functions_service::wasm::api::StdWasmApiImpl

pub struct StdWasmApiImpl<const S: usize> { /* private fields */ }

Implementation of the standard Oak Functions API.

There are probably more locks than necessary here, it should be possible to reduce them in the future.